refactor(model): consolidate lifecycle tests + document breaking changes
Phase 2 (AC-002) + Phase 3 (AC-003) of #93.

Phase 2: extract three duplicated lifecycle test bodies from
embed/tests.rs and reranker/tests.rs into generic helpers in
test_support.rs:

- assert_cache_lookup_returns_some_when_all_files_present<Id>
- assert_cache_lookup_returns_none_when_empty<Id>
- assert_from_probe_error_maps_correctly (kind-agnostic, since
  ModelInitError is unified across embed and reranker)

Each per-kind test now invokes the helper in a single line,
eliminating the parallel-duplicated bodies (~40 lines total).

Phase 3: add four breaking-change bullets to CHANGELOG.md covering the
ModelInitError unification, download_model/cached_artifacts signature
change, ModelArtifact::Kind associated type, and CandidateArtifacts
generic.

Validates T-018, T-019, T-020, T-021, T-022, T-025.
cargo test --workspace: 308 passed / 0 failed / 3 ignored.
cargo clippy --workspace --all-targets -- -D warnings: clean.
cargo fmt --all -- --check: clean.
